AS-5077

+++ src/main/java/com/asg/shipchandling/salesquotationsch/dto/request/CreateSalesQuotationSchRequest.java
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -27,7 +28,7 @@ public class CreateSalesQuotationSchRequest {
     @Size(max = 20, message = "Currency code must not exceed 20 characters")
     private String currencyCode;
 
-    private Long currencyRate;
+    private BigDecimal currencyRate;
 
     @Size(max = 20, message = "Quotation status must not exceed 20 characters")
     private String quotationStatus;
