+++ src/main/java/com/asg/shipchandling/salesquotationsch/repository/SalesQuotationSchStoredProcRepository.java
@@ -52,7 +52,7 @@ public class SalesQuotationSchStoredProcRepository {
                             String currencyCode = rs.getString("CURRENCY_CODE");
                             dto.setCurrencyCode(currencyCode);
 
-                            Long currencyRate = rs.getLong("CURRENCY_RATE");
+                            BigDecimal currencyRate = rs.getBigDecimal("CURRENCY_RATE");
                             dto.setCurrencyRate(currencyRate);
 
                             String paymentMode = rs.getString("PAYMENT_MODE");
